+++ grails-app/controllers/zelifkudos/AuthInterceptor.groovy
@@ -3,7 +3,7 @@ package zelifkudos
 class AuthInterceptor {
 
     AuthInterceptor() {
-        matchAll().excludes(controller: 'login').excludes(controller: 'pwa').excludes(uri: '/ws/**')
+        matchAll().excludes(controller: 'login').excludes(controller: 'pwa').excludes(controller: 'demo').excludes(uri: '/ws/**')
     }
 
     boolean before() {